Re #336: Transaction nonce now in terms of BigInteger
- Both Transaction and InternalTransaction take the nonce input
  as a BigInteger instead of a long. This is to sync this class up
  with the KernelInterface using BigInteger, which had previously
  caused us to have to extract the long type out of the BigInteger
  in nearly every test.

+++ org.aion.avm.core/src/org/aion/avm/core/types/InternalTransaction.java
@@ -12,7 +12,7 @@ public class InternalTransaction extends Transaction implements InternalTransact
 
     private boolean rejected ;
 
-    public InternalTransaction(Type type, Address from, Address to, long nonce, BigInteger value, byte[] data, long energyLimit, long energyPrice) {
+    public InternalTransaction(Type type, Address from, Address to, BigInteger nonce, BigInteger value, byte[] data, long energyLimit, long energyPrice) {
         super(type, from, to, nonce, value, data, energyLimit, energyPrice);
     }
