+++ org.aion.avm.core/test/org/aion/avm/core/AvmFailureTest.java
@@ -46,7 +46,7 @@ public class AvmFailureTest {
         
         byte[] jar = JarBuilder.buildJarForMainAndClasses(AvmFailureTestResource.class);
         byte[] arguments = null;
-        Transaction tx = Transaction.create(deployer, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, new CodeAndArguments(jar, arguments).encodeToBytes(), energyLimit, energyPrice);
+        Transaction tx = Transaction.create(deployer, kernel.getNonce(deployer), BigInteger.ZERO, new CodeAndArguments(jar, arguments).encodeToBytes(), energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         TransactionResult txResult = avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -62,7 +62,7 @@ public class AvmFailureTest {
     @Test
     public void testFailedTransaction() {
         byte[] data = ABIEncoder.encodeMethodArguments("reentrantCall", 5);
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         AvmTransactionResult txResult = (AvmTransactionResult) avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -78,7 +78,7 @@ public class AvmFailureTest {
     @Test
     public void testOutOfEnergy() {
         byte[] data = ABIEncoder.encodeMethodArguments("testOutOfEnergy");
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         TransactionResult txResult = avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -88,7 +88,7 @@ public class AvmFailureTest {
     @Test
     public void testOutOfStack() {
         byte[] data = ABIEncoder.encodeMethodArguments("testOutOfStack");
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         TransactionResult txResult = avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -98,7 +98,7 @@ public class AvmFailureTest {
     @Test
     public void testRevert() {
         byte[] data = ABIEncoder.encodeMethodArguments("testRevert");
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         AvmTransactionResult txResult = (AvmTransactionResult) avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -110,7 +110,7 @@ public class AvmFailureTest {
     @Test
     public void testInvalid() {
         byte[] data = ABIEncoder.encodeMethodArguments("testInvalid");
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         AvmTransactionResult txResult = (AvmTransactionResult) avm.run(new TransactionContext[] {txContext})[0].get();
 
@@ -122,7 +122,7 @@ public class AvmFailureTest {
     @Test
     public void testUncaughtException() {
         byte[] data = ABIEncoder.encodeMethodArguments("testUncaughtException");
-        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer).longValue(), BigInteger.ZERO, data, energyLimit, energyPrice);
+        Transaction tx = Transaction.call(deployer, dappAddress, kernel.getNonce(deployer), BigInteger.ZERO, data, energyLimit, energyPrice);
         TransactionContext txContext = new TransactionContextImpl(tx, block);
         AvmTransactionResult txResult = (AvmTransactionResult) avm.run(new TransactionContext[] {txContext})[0].get();
